requirements.txt
bs4==0.0.1
h5py==3.8.0
keras==2.11.0
matplotlib==3.5.1
numpy==1.24.2
pandas==1.3.5
Pillow==8.2.0
scikit-learn==1.2.2
scipy==1.10.1
sklearn==0.0.post1
spectral==0.23.1
streamlit==1.18.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
